Refactor to use Waypoint manager
