+++ src/main/java/io/github/lichen911/waypoints/Waypoints.java
@@ -5,6 +5,7 @@ import org.bukkit.plugin.java.JavaPlugin;
 
 import io.github.lichen911.waypoints.commands.WpCommand;
 import io.github.lichen911.waypoints.managers.WaypointManager;
+import io.github.lichen911.waypoints.objects.CommandLiteral;
 import io.github.lichen911.waypoints.utils.ConfigReader;
 
 public final class Waypoints extends JavaPlugin {
